pyproject.toml
[project]
name = "resume-online"
version = "1.0.0"
description = "Interactive resume generator with PDF export"
requires-python = ">=3.11"
dependencies = [
    "jinja2>=3.1.4",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "weasyprint>=62.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

